Refactor main.py, exception handling
Refactored load_config() and split it into 3 functions, one function for each file accessed. Each function now returns the data

Global statements are now in on_ready() and so are the global assignments

Globals start initialized with type hinting instead of having fallback data that was never utilized

Added type hinting to the message argument in on_message() for easier access to the data type's members

Finished code that logged a message to a user's DMs and turned it into a function called send_log_to_dm() as it was used on more than on occasion

Added a check in on_ready() that checks if all privileged intents are enabled. If not, then an error message prints and the program exits